requirements.txt
# Core / Testing
garth>=0.4.44
python-dotenv>=1.0.1
pandas>=2.2.0
matplotlib>=3.8.3
pytest>=8.0.1

# Flask extensions
Flask>=3.0.3
Flask-SQLAlchemy>=3.1.1
Flask-WTF>=1.2.1
Flask-Session>=0.8.0

# Debug
Flask-DebugToolbar>=0.14.1